Add captures and scripts to reproduce them
As a way of allowing automated testing, we capture the output
of the tablet from hid-record while using it. Later, we can replay
that sequence of events through hid-replay and make sure that the
kernel events (recorded from evemu-record) and X11 events (recorded
from xinput test-xi2) do not change.

This is a quickly hacked-together sample, so use at your own risk :P