scripts: Add a 'sysinfo.sh' script which dumps system/tablet information
It can be hard to explain exactly what a normal user should do to get
useful hardware information from their system. This new script is trivial
to run and gathers a a great deal of relevant data into a tarball at
/tmp/sysinfo.??????????.tar.gz which can be sent over email or uploaded
for retrieval.

This script does not require root permissions to run.