scripts: Improve functionality of sysinfo.sh script
Multiple improvements to the sysinfo.sh script, including
improved error handling and additional gathered data. In
addition, it detects if it is run as root and gatheres additional
data where possible (e.g. getting the HID descriptors on pre-3.17
kernels requires us to temporarily unload the kernel driver, and
some distros prevent us from using the udevadm command without
root). As such, it now requests to be run as root, though it will
still gather what data it can as a regular user.