scripts: Capture a sysfs tree of Wacom devices from sysinfo
If there are Wacom devices which aren't bound as expected or otherwise
missed by other sections in this script, it would be nice to at least
have them mentioned somewhere in the sysinfo output. This commit has
the script scan through the sysfs tree for potential Wacom devices and
produce a tree-like output of the filesystem without regard for what
the nodes are.