scripts: Fix 'libinput-list-devices' call on newer systems
Newer versions of libinput have moved away from providing a number of
separate utility commands to having everyting available through just
the 'libinput' command. This breaks the sysinfo script's call out to
libinput-list-devices. This commit now first checks if the command
exists and runs either it or the new alternative.

Ref: https://github.com/linuxwacom/wacom-hid-descriptors/issues/37